fix: Resolve gap problem
In the previous commit, there is a problem that if there are no trading
days like 2023-12-22 to 2023-12-26.
In that case, the market will be closed in the weekend and on Monday
due to holiday. However, when we plot the graph,
there are gaps and lines in the graph, which we would not expect.
This commit is for that problem.
Additionally, to plot the graph, an additional decorator is used for
the x-axis format.

1. Use decorator to validate the datetime whether it is traded
on the opening day of the market.
2. Plot the graphs without any gap
3. Use decorator for time format of x-axis
4. Use __call__ method for error handling